Rescale and remove sigmoid from regression model
I have removed the sigmoid activation from the final layer of the
regression model. This was necessary because the sigmoid saturates at
-1 and 1. Often, more extreme output is required.

Moreover, I have changed the scaling from dividing and multiplying by
the variance to dividing and multiplying by the standard deviation. I
don't know why I did not do this earlier.